fixed warnings for #396

+++ SpotHopper/ClientSessionManager.m
@@ -75,14 +75,14 @@
 - (AFHTTPRequestOperation *)GET:(NSString *)URLString parameters:(NSDictionary *)parameters success:(void (^)(AFHTTPRequestOperation *, id))success {
     return [super GET:URLString parameters:parameters success:^(AFHTTPRequestOperation *operation, id responseObject) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
         
         success(operation, responseObject);
     } failure:^(AFHTTPRequestOperation *operation, NSError *error) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
         
@@ -101,7 +101,7 @@
         if (_debug) {
             NSLog(@"Request Headers\n\t%@", operation.request.allHTTPHeaderFields);
             NSLog(@"Request Body\n\t%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
-            NSLog(@"Response\n\t%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"Response\n\t%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
         }
         
         success(operation, responseObject);
@@ -109,7 +109,7 @@
         if (_debug) {
             NSLog(@"Request Headers\n\t%@", operation.request.allHTTPHeaderFields);
             NSLog(@"Reques Body\n\t%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
-            NSLog(@"Response\n\t%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"Response\n\t%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
         }
         
         id response = nil;
@@ -125,7 +125,7 @@
 - (AFHTTPRequestOperation *)POST:(NSString *)URLString parameters:(NSDictionary *)parameters success:(void (^)(AFHTTPRequestOperation *, id))success {
     return [super POST:URLString parameters:parameters success:^(AFHTTPRequestOperation *operation, id responseObject) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"Request Headers - %@", operation.request.allHTTPHeaderFields);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
@@ -137,7 +137,7 @@
         if (_debug) {
             NSLog(@"Request Headers - %@", operation.request.allHTTPHeaderFields);
             NSLog(@"Request Body - %@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
-            NSLog(@"Response - %@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"Response - %@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
         }
         
         id response = nil;
@@ -153,14 +153,14 @@
 - (AFHTTPRequestOperation *)PUT:(NSString *)URLString parameters:(NSDictionary *)parameters success:(void (^)(AFHTTPRequestOperation *, id))success {
     return [super PUT:URLString parameters:parameters success:^(AFHTTPRequestOperation *operation, id responseObject) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
         
         success(operation, responseObject);
     } failure:^(AFHTTPRequestOperation *operation, NSError *error) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
         
@@ -177,14 +177,14 @@
 - (AFHTTPRequestOperation *)DELETE:(NSString *)URLString parameters:(NSDictionary *)parameters success:(void (^)(AFHTTPRequestOperation *, id))success {
     return [super DELETE:URLString parameters:parameters success:^(AFHTTPRequestOperation *operation, id responseObject) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
         
         success(operation, responseObject);
     } failure:^(AFHTTPRequestOperation *operation, NSError *error) {
         if (_debug) {
-            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, operation.response.statusCode, operation.responseString);
+            NSLog(@"%@ %ld - %@", operation.request.URL.standardizedURL, (long)operation.response.statusCode, operation.responseString);
             NSLog(@"%@", [[NSString alloc] initWithData:operation.request.HTTPBody encoding:NSUTF8StringEncoding]);
         }
